Move getAuthorNick from TicketMessageRepository to TicketService
